Fixes to enable larger traces and avoid memory leaks

+++ src/app/inc/rocprofvis_structs.h
@@ -25,8 +25,8 @@ typedef struct rocprofvis_color_by_value_t
 
 typedef struct rocprofvis_data_point_t
 {
-    float xValue;
-    float yValue;
+    double xValue;
+    double yValue;
 } rocprofvis_data_point_t;
 
 typedef struct rocprofvis_graph_map_t
@@ -70,9 +70,9 @@ struct rocprofvis_trace_process_t
 
 struct rocprofvis_meta_map_struct_t
 {
-    float       size;
-    float       max;
-    float       min;
+    double       size;
+    double       max;
+    double       min;
     std::string type;
     std::string chart_name;
 };
