+++ src/view/src/rocprofvis_flame_chart.h
@@ -16,17 +16,17 @@ namespace View
 class FlameChart : public Charts
 {
 public:
-    FlameChart(int chart_id, std::string name, float zoom, float movement, float min_x,
-               float max_x, float scale_x);
+    FlameChart(int chart_id, std::string name, double zoom, double movement, double min_x,
+               double max_x, double scale_x);
     void Render() override;
     void DrawBox(ImVec2 start_position, int boxplot_box_id,
-                 rocprofvis_trace_event_t flame, float duration, ImDrawList* draw_list);
+                 rocprofvis_trace_event_t flame, double duration, ImDrawList* draw_list);
     void ExtractFlamePoints(rocprofvis_controller_array_t* track_data);
 
-    std::tuple<float, float> FindMaxMinFlame();
-    void        UpdateMovement(float zoom, float movement, float& min_x, float& max_x,
-                               float scale_x) override;
-    float       ReturnSize() override;
+    std::tuple<double, double> FindMaxMinFlame();
+    void        UpdateMovement(double zoom, double movement, double& min_x, double& max_x,
+                               double scale_x) override;
+    double       ReturnSize() override;
     int         ReturnChartID() override;
     void        SetID(int id) override;
     std::string GetName() override;
@@ -35,16 +35,16 @@ public:
 private:
     std::vector<rocprofvis_trace_event_t> flames;
     std::string                           m_name;
-    float                                 m_min_value;
-    float                                 m_max_value;
-    float                                 m_min_x;
-    float                                 m_max_x;
-    float                                 m_zoom;
-    float                                 m_movement;
-    float                                 m_scale_x;
+    double                                 m_min_value;
+    double                                 m_max_value;
+    double                                 m_min_x;
+    double                                 m_max_x;
+    double                                 m_zoom;
+    double                                 m_movement;
+    double                                 m_scale_x;
     int                                   m_chart_id;
-    float                                 size;
-    float                                 m_sidebar_size;
+    double                                 size;
+    double                                 m_sidebar_size;
     rocprofvis_color_by_value_t             color_by_value_digits;
 };
