+++ src/view/src/rocprofvis_grid.h
@@ -15,8 +15,8 @@ public:
     Grid();
     ~Grid();
 
-    void RenderGrid(float min_x, float max_x, float movement, float zoom,
-                    ImDrawList* draw_list, float scale_x, float v_max_x, float v_min_x);
+    void RenderGrid(double min_x, double max_x, double movement, double zoom,
+                    ImDrawList* draw_list, double scale_x, double v_max_x, double v_min_x);
 };
 
 }  // namespace View
