+++ src/view/src/rocprofvis_line_chart.h
@@ -18,40 +18,40 @@ namespace View
 class LineChart : public Charts
 {
 public:
-    LineChart(int id, std::string name, float zoom, float movement, float& min_x,
-              float& max_x, float scale_x);
+    LineChart(int id, std::string name, double zoom, double movement, double& min_x,
+              double& max_x, double scale_x);
     ~LineChart();
     void   Render() override;
-    void   UpdateMovement(float zoom, float movement, float& min_x, float& max_x,
-                          float scale_x) override;
+    void   UpdateMovement(double zoom, double movement, double& min_x, double& max_x,
+                          double scale_x) override;
     ImVec2 MapToUI(rocprofvis_data_point_t& point, ImVec2& c_position, ImVec2& c_size,
-                   float scale_x, float scale_y);
+                   double scale_x, double scale_y);
     std::vector<rocprofvis_data_point_t> LineChart::ExtractPointsFromData(
         rocprofvis_controller_array_t* track_data);
-    std::tuple<float, float> FindMaxMin();
-    float                    ReturnSize() override;
+    std::tuple<double, double> FindMaxMin();
+    double                    ReturnSize() override;
     void                     SetID(int id) override;
     int                      ReturnChartID() override;
     std::string              GetName() override;
     int                      SetSize();
     void  SetColorByValue(rocprofvis_color_by_value_t color_by_value_digits) override;
-    float CalculateMissingX(float x1, float y1, float x2, float y2, float known_y);
+    double CalculateMissingX(double x1, double y1, double x2, double y2, double known_y);
 
 private:
     std::vector<rocprofvis_data_point_t>   m_data;
     std::string                            m_name;
-    float                                  m_min_value;
-    float                                  m_max_value;
-    float                                  m_zoom;
+    double                                  m_min_value;
+    double                                  m_max_value;
+    double                                  m_zoom;
     rocprofvis_color_by_value_t              m_color_by_value_digits;
-    float                                  m_movement;
-    float                                  m_min_x;
-    float                                  m_max_x;
-    float                                  m_min_y;
-    float                                  m_max_y;
-    float                                  m_scale_x;
+    double                                  m_movement;
+    double                                  m_min_x;
+    double                                  m_max_x;
+    double                                  m_min_y;
+    double                                  m_max_y;
+    double                                  m_scale_x;
     int                                    m_id;
-    float                                  size;
+    double                                  size;
     bool                                   is_color_value_existant;
     std::map<int, rocprofvis_graph_map_t>* tree;
 };
